refactor(#4794): extract race-loss requeue tests
Keep the namespace below its size cap without changing production behavior or test assertions.
